# fix: optimize codegen asset upload